Consolidate `product` handlers into a single file, refactor request structs, and improve test consistency
