--- internal/product/delete.go
@@ -1,30 +1,0 @@
-package product
-
-import (
-	"context"
-	"errors"
-	"net/http"
-	"products/internal"
-	"time"
-
-	"github.com/jackc/pgx/v5"
-)
-
-func (h *productHandler) DeleteProduct(w http.ResponseWriter, r *http.Request) {
-	id, ok := internal.GetIntFromRequestPath("id", r)
-	if !ok {
-		http.Error(w, "Invalid product ID", http.StatusBadRequest)
-		return
-	}
-	contextWithTimeOut, cancel := context.WithTimeout(r.Context(), 5*time.Second)
-	defer cancel()
-	if _, err := h.queries.DeleteProduct(contextWithTimeOut, id); err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			http.Error(w, "Product not found", http.StatusNotFound)
-			return
-		}
-		http.Error(w, "Failed to delete product", http.StatusInternalServerError)
-		return
-	}
-	w.WriteHeader(http.StatusNoContent)
-}
